+++ lib/bali.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'bali/version'
-require 'bali/engine'
 require 'bali/filter_form'
 require 'bali/form_builder/html_utils'
 require 'bali/form_builder/shared_utils'
@@ -16,6 +14,7 @@ require 'bali/path_helper'
 require 'bali/form_helper'
 require 'bali/auto_submit_select_helper'
 require 'bali/icon_tag_helper'
+require 'bali/time_periods/select_options'
 
 Dir[File.join(File.dirname(__FILE__), 'bali/concerns', '**/*.rb')].each do |concern|
   require concern
@@ -32,6 +31,9 @@ Dir[File.join(File.dirname(__FILE__), 'bali/commands', '**/*.rb')].each do |comm
 end
 
 require 'bali/form_builder'
+
+require 'bali/version'
+require 'bali/engine'
 
 module Bali
   mattr_accessor :native_app, default: false