+++ spec/dummy/app/javascript/application.js
@@ -26,6 +26,7 @@ import { SubmitButtonController } from 'bali/submit-button-controller'
 import { SubmitOnChangeController } from 'bali/submit-on-change-controller'
 import { TrixAttachmentsController } from 'bali/trix-attachments-controller'
 import { RecurrentEventRuleController } from 'bali/recurrent-event-rule'
+import { TimePeriodFieldController } from 'bali/time-period-field-controller'
 
 import { AvatarController } from 'bali/avatar'
 import { BulkActionsController } from 'bali/bulk_actions'
@@ -110,5 +111,6 @@ application.register('slim-select', SlimSelectController)
 application.register('step-number-input', StepNumberInputController)
 application.register('submit-button', SubmitButtonController)
 application.register('submit-on-change', SubmitOnChangeController)
+application.register('time-period-field', TimePeriodFieldController)
 application.register('trix-attachments', TrixAttachmentsController)
 application.register('recurrent-event-rule', RecurrentEventRuleController)