+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.4.7] - 2025-09-29
+## [1.4.8] - 2025-09-04
+
+### Changed
+
+- `time_period_select_field` and `time_period_select_field_group` to `Bali::FormBuilder`
+- `Bali::TimePeriods::SelectOptions` as default time periods for `time_period_select_*` fields.
+
+## [1.4.7] - 2025-08-29
 
 ### Changed